fix(format): address SARIF review notes
